+++ src/states/game/StartMenu.lua
@@ -4,11 +4,21 @@ function StartMenu:init()
     self.type = 'StartMenu'
 
     self.playButton = Button(BUTTON_PARAMS['Play'])
-    self.backgroundFrame = gFrames['StartMenuBackground']
+    self.newButton = Button(BUTTON_PARAMS['New'])
+    self.background = StartMenuBackground()
+
+    if love.filesystem.getInfo(SAVE_FILE) then
+        self.playButton:enable()
+    end
+
     self.interactables = {
         self.playButton,
+        self.newButton,
     }
+
+    gStateStack:push(self.background)
     gStateStack:push(self.playButton)
+    gStateStack:push(self.newButton)
 end
 
 function StartMenu:update(dt)
@@ -16,10 +26,4 @@ function StartMenu:update(dt)
 end
 
 function StartMenu:render()
-    --[[Font and text 
-    love.graphics.setFont(gFonts['large'])
-    love.graphics.printf('Start Menu', 0, VIRTUAL_HEIGHT / 2 - 6, VIRTUAL_WIDTH, 'center')]]
-    love.graphics.draw(self.backgroundFrame, 0, 0, 0, 
-        VIRTUAL_WIDTH / self.backgroundFrame:getWidth(), VIRTUAL_HEIGHT / self.backgroundFrame:getHeight())
-    self.playButton:render()
 end